Replace floats with Decimal in tablegenerator
Decimal has many advantages compared to floats such as support for
correct rounding and the ability to precisely represent decimal numbers.

Decimal was already used for statistics calculations, now we also use it
when rounding and formatting values.

This commit contains the minimal amount of changes necessary to make
this work, without any further fixes and simplifications.

Due to the increased precision of Decimal, there are now also
differences in the output of our integration tests,
many of which are actually improvements.
To keep the commit size small, these will be tackled in subsequent commits,
so the table-generator integration tests are expected to fail now.

+++ benchexec/tablegenerator/test_columns.py
@@ -5,6 +5,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+from decimal import Decimal
 import unittest
 
 from benchexec.tablegenerator.columns import (
@@ -14,8 +15,8 @@ from benchexec.tablegenerator.columns import (
 )
 from benchexec.tablegenerator.util import TableDefinitionError
 
-nan = float("nan")
-inf = float("inf")
+nan = Decimal("nan")
+inf = Decimal("inf")
 
 
 class ColumnsTest(unittest.TestCase):
